Fixing playback with gstreamer
When contenttype is empty/not available, MediaPlayer::loadWithNextMediaEngine
is using nextMediaEngine method to select the media engine.
nextMediaEngine basically returns the first media engine from the media engines vector.
Currently, if WebRTC or hole punch gstreamer is enabled,
it will be selected and the playback will fail.
The patch fixes this by setting network to failed state for inappropriate engines.